Added EventBridge scheduler to invoke pipeline every day at 10 PM
